+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * **New** - Elementor Dynamic Tag: Criteo Product ID
 * **New** - ACF for Discovery: new field video
+* **New** - Post class for CPT without urls, to allow hiding parts of the archive page (loop)
 
 ### 1.1.7: September 20th, 2018
 * **New** - Elementor Dynamic Tag: Resaweb URL No Accommodation